decouple OrganizeMap from Communities section
